Crate local_ip_address

Local IP Address

Retrieve system’s local IP address and Network Interfaces/Adapters on Linux, macOS and Windows.

Usage

Get the local IP address of your system by executing the local_ip function:

use local_ip_address::local_ip;

let my_local_ip = local_ip().unwrap();

println!("This is my local IP address: {:?}", my_local_ip);

Retrieve all the available network interfaces from both, the AF_INET and the AF_INET6 family by executing the list_afinet_netifas function:

use local_ip_address::list_afinet_netifas;

let network_interfaces = list_afinet_netifas().unwrap();

for (name, ip) in network_interfaces.iter() {
    println!("{}:\t{:?}", name, ip);
}

Underlying approach on retrieving network interfaces or the local IP address may differ based on the running operative system.

OS	Approach

Linux	Establishes a Netlink socket interchange to retrieve network interfaces

macOS	Uses of getifaddrs to retrieve network interfaces

Windows	Consumes Win32 API’s to retrieve the network adapters table

Re-exports

pub use crate::macos::*;

Modules

macos

Enums

Error

Functions

find_ifa

Finds the network interface with the provided name in the vector of network interfaces provided

local_ip

Retrieves the local ip address of the machine in the local network from the AF_INET family.