Module local_ip_address::linux

Functions

list_afinet_netifas

Perform a search over the system’s network interfaces using Netlink Route information, retrieved network interfaces belonging to both socket address families AF_INET and AF_INET6 are retrieved along with the interface address name.

local_ip

Retrieves the local IPv4 address for this system

local_ipv6

Retrieves the local IPv6 address for this system